update to plots.py
This (hopefully) shows how I generate the combined plots. Its sort of manual but I think as it combines many different experiments this is the fastest way. Note that we have to pad the length of the lists of the experiments with the last value such that they are all the same length (if this doesnt make sense ask me).